+++ src/components/Modal/Modal.js
@@ -43,7 +43,7 @@ export const Modal = ({
           <>
             <div
               onClick={setToggle.bind(this, false)}
-              className={`fixed left-0per top-0per bottom-0per right-0per backdrop-blur-30px text-black ${backgroundAnimation}`}
+              className={`fixed z-97 left-0per top-0per bottom-0per right-0per backdrop-blur-30px text-black ${backgroundAnimation}`}
               style={{ backgroundColor: 'rgba(16,16,16,0.5)' }}
             />
             <ModalContent
@@ -70,7 +70,7 @@ const ModalContent = ({ children, className, modalContentAnimation, closeIcon, s
       ref={modalRef}
       tabIndex={0}
       {...restProps}
-      className={`text-black overflow-x-hidden fixed left-50per top-50per rounded-20px bg-white min-w-50rem max-w-50vw sm:min-w-90vw md:min-w-90vw max-h-75vh ${
+      className={`fixed left-50per top-50per z-98 text-black overflow-x-hidden rounded-20px bg-white min-w-50rem max-w-50vw sm:min-w-90vw md:min-w-90vw max-h-75vh ${
         shadow ? 'shadow-small-black-2' : ''
       } ${className} ${modalContentAnimation}`}
       style={{ transform: 'translate(-50%, -50%)' }}>