+++ purgecss.config.js
@@ -3,6 +3,26 @@ module.exports = {
   css: ['./node_modules/greencss/css/minified/greencss.css'],
   keyframes: true,
   defaultExtractor: (content) => content.match(/[A-Za-z0-9-_:/]+/g) || [],
-  blocklist: ['html', 'body', 'div', 'header', 'h3', 'p', 'button', 'svg', 'footer', 'span'],
+  blocklist: [
+    'html',
+    'body',
+    'div',
+    'header',
+    'h3',
+    'p',
+    'button',
+    'svg',
+    'footer',
+    'span',
+    'a',
+    'i',
+    'center',
+    'label',
+    'table',
+    'canvas',
+    '*',
+    'container',
+    'sans'
+  ],
   output: './src/style/green.css'
 }