+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 ## [7.0.2](https://github.com/Se-Gl/codn/compare/v7.0.1...v7.0.2) (2022-08-02)
 
-
 ### Bug Fixes
 
-* **modal:** adjusted modal opacity ([966a973](https://github.com/Se-Gl/codn/commit/966a973b679a9a2c0efc9c587ef3ac26cd9ab34a))
+- **modal:** adjusted modal opacity ([966a973](https://github.com/Se-Gl/codn/commit/966a973b679a9a2c0efc9c587ef3ac26cd9ab34a))
 
 ## [7.0.1](https://github.com/Se-Gl/codn/compare/v7.0.0...v7.0.1) (2022-08-02)